+++ R/deslant.R
@@ -3,10 +3,10 @@
 #' @description This algorithm sets handwritten text in images upright by removing cursive writing style. 
 #' One can use it as a preprocessing step for handwritten text recognition.\cr
 #' \itemize{
-#' \item{ocv_deslant expects a magick-image and does not performs grayscaling}
-#' \item{image_deslant expects a magick-image and performs grayscaling when calling the function}
+#' \item{\code{image_deslant} expects a magick-image and performs grayscaling before doing deslanting}
+#' \item{\code{ocv_deslant} expects a ocv-image and does not perform grayscaling before doing deslanting}
 #' }
-#' @param image an object of class opencv-image (ocv_deslant) with pixel values between 0 and 255 or a magick-image (image_deslant)
+#' @param image an object of class opencv-image (for \code{ocv_deslant}) with pixel values between 0 and 255 or a magick-image (for \code{image_deslant})
 #' @param bgcolor integer value with the background color to use to fill the gaps of the sheared image that is returned. Defaults to white: 255
 #' @param lower_bound lower bound of shear values. Defaults to -1
 #' @param upper_bound upper bound of shear values. Defaults to 1
@@ -27,6 +27,7 @@
 #' up
 #' 
 #' img  <- image_read(path)
+#' img
 #' image_deslant(img)
 #' \dontshow{
 #' \}
@@ -35,14 +36,14 @@
 ocv_deslant <- function(image, bgcolor = 255, lower_bound = -1, upper_bound = 1){
   stopifnot(inherits(image, "opencv-image"))
   bgcolor <- as.integer(bgcolor)
-  image.textlinedetector:::textlinedetector_deslant(image, bgcolor, lower_bound, upper_bound)
+  textlinedetector_deslant(image, bgcolor, lower_bound, upper_bound)
 }
 
 #' @export
 #' @rdname ocv_deslant
-image_deslant <- function(image, bgcolor = 255, lower_bound = -1, upper_bound = 1, ...){
+image_deslant <- function(image, bgcolor = 255, lower_bound = -1, upper_bound = 1){
   stopifnot(inherits(image, "magick-image"))
-  img <- magick_to_opencv(image, ...)
+  img <- magick_to_opencv(image)
   img <- opencv::ocv_grayscale(img)
   deslanted <- ocv_deslant(img, bgcolor, lower_bound, upper_bound)
   opencv_to_magick(deslanted)